backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
astral
geographiclib
python-dotenv
requests
ephem
